pkg/schedgen: Fix team 1 issues.
Solve insideous problem with schedule generation not including team 1.
This was related to various places where numbers are being translated
from code space (0-indexed) to database space (1-indexed).  This is
error proned and just doesn't work well when we consider that the
schedule is only ever used to populate the database which then
operates in database-space from that point forward.  To reconcile
this, we just operate entirely in code-space in the schedule
generator, and then in the handler that indexes the selected team
list, we transparently convert to database-space.

Additionally, this commit adds some validation components to make sure
that all teams appear that were expected to and that all teams play
the same number of rounds.